requirements.txt
Django==3.1.7
djangorestframework==3.12.4
pytz==2021.1
mysqlclient==2.1.0
django-cors-headers==3.7.0